Revert "Merge branch 'gtk' into gtk"
This reverts commit 7ac53ed8915456e5083969277d337fa5c2865bdb, reversing
changes made to 52c63a4763e02787f105395ab4de0684b9a906aa.
