+++ manuskript/data/abstractData.py
@@ -20,11 +20,6 @@ class AbstractData:
         self.dataPath = path
         self.dataStatus = DataStatus.UNDEFINED
 
-    def changePath(self, path: str):
-        print("{} -> {}".format(self.dataPath, path))
-
-        self.dataPath = path
-
     def complete(self, statusCompletion: bool = True):
         if self.dataStatus == DataStatus.LOADING:
             self.dataStatus = DataStatus.LOADED if statusCompletion else DataStatus.UNDEFINED
