+++ manuskript/ui/util.py
@@ -8,7 +8,6 @@ gi.require_version('GdkPixbuf', '2.0')
 from gi.repository import GdkPixbuf, Gdk
 
 from manuskript.data import Color, OutlineItem, OutlineText, OutlineFolder
-from manuskript.util import profileTime
 
 
 def rgbaFromColor(color: Color) -> Gdk.RGBA:
@@ -33,36 +32,6 @@ def bindMenuItem(builder, id, action):
         return
 
     menuItem.connect("activate", action)
-
-
-def packViewIntoSlot(slot, view_cls, data=None):
-    if slot is None:
-        return None
-
-    for child in slot.get_children():
-        slot.remove(child)
-
-    try:
-        if data is None:
-            view = profileTime(view_cls)
-        else:
-            view = profileTime(view_cls, data)
-    except Exception:
-        return None
-
-    if view.widget is None:
-        return None
-
-    slot.pack_start(view.widget, True, True, 0)
-    return view
-
-
-def unpackFromSlot(slot, view):
-    if (slot is not None) and (view.widget is not None):
-        slot.remove(view.widget)
-
-    del view
-    return None
 
 
 def iconByOutlineItemType(outlineItem: OutlineItem) -> str:
