+++ manuskript/util/__init__.py
@@ -4,7 +4,6 @@
 import re
 import time
 import traceback
-import urllib.parse
 
 from manuskript.util.counter import CounterKind, CharCounter, WordCounter, PageCounter
 
@@ -50,18 +49,6 @@ def safeFilename(filename: str, extension: str = None) -> str:
         name = "_" + name
 
     return re.sub(r"[^a-zA-Z0-9._\-+()]", "_", name)
-
-
-def parseFilenameFromURL(url: str) -> str | None:
-    result = urllib.parse.urlparse(url)
-
-    if result is None:
-        return None
-
-    if result.scheme != "file":
-        return None
-
-    return result.path
 
 
 def countText(text: str, kind: CounterKind = CounterKind.WORDS):
