+++ ui/startup.glade
@@ -8,7 +8,7 @@
       <pattern>*.msk</pattern>
     </patterns>
   </object>
-  <object class="GtkRecentChooserMenu" id="recent_chooser_menu_btn">
+  <object class="GtkRecentChooserMenu" id="recent_chooser_menu">
     <property name="visible">True</property>
     <property name="can-focus">False</property>
     <property name="filter">recent_filter</property>
@@ -91,10 +91,9 @@
                         <property name="label" translatable="yes">_Recent</property>
                         <property name="use-underline">True</property>
                         <child type="submenu">
-                          <object class="GtkRecentChooserMenu" id="recent_chooser_menu">
+                          <object class="GtkRecentChooserMenu" id="recent_chooser_menu1">
                             <property name="visible">True</property>
                             <property name="can-focus">False</property>
-                            <property name="filter">recent_filter</property>
                             <property name="limit">10</property>
                             <property name="sort-type">mru</property>
                           </object>
@@ -458,7 +457,7 @@
                 <property name="halign">end</property>
                 <property name="spacing">9</property>
                 <child>
-                  <object class="GtkButton" id="open_button">
+                  <object class="GtkButton">
                     <property name="visible">True</property>
                     <property name="can-focus">True</property>
                     <property name="receives-default">True</property>
@@ -501,12 +500,12 @@
                   </packing>
                 </child>
                 <child>
-                  <object class="GtkMenuButton" id="recent_button">
+                  <object class="GtkMenuButton">
                     <property name="visible">True</property>
                     <property name="can-focus">True</property>
                     <property name="focus-on-click">False</property>
                     <property name="receives-default">True</property>
-                    <property name="popup">recent_chooser_menu_btn</property>
+                    <property name="popup">recent_chooser_menu</property>
                     <property name="use-popover">False</property>
                     <child>
                       <object class="GtkBox">
@@ -547,7 +546,7 @@
                   </packing>
                 </child>
                 <child>
-                  <object class="GtkButton" id="create_button">
+                  <object class="GtkButton">
                     <property name="visible">True</property>
                     <property name="can-focus">True</property>
                     <property name="receives-default">True</property>
